WIP: refactors day 1 and day 2
