+++ challenges/day_01/challenge_01.py
@@ -1,9 +1,14 @@
+import sys
 import os
 
-if __name__ == "__main__":
-    from lib.calculate_fuel import calculate_total_fuel_from_list
-else:
-    from challenges.day_01.lib.calculate_fuel import calculate_total_fuel_from_list
+
+sys.path.append(os.path.join(os.path.dirname(
+    os.path.dirname(sys.path[0]))))
+
+try:
+    from lib.calculate_fuel.calculate_fuel import calculate_total_fuel_from_list
+except:
+    raise
 
 
 def day_01_challenge_01():
