+++ lib/calculate_fuel/calculate_fuel.py
