+++ lib/calculate_fuel/tests/test_calculate_fuel_compounded.py
@@ -1,7 +1,7 @@
 import pytest
 
-from challenges.day_01.lib.calculate_fuel import *
-from challenges.day_01.lib.calculate_fuel_compounded import *
+from lib.calculate_fuel.calculate_fuel import *
+from lib.calculate_fuel.calculate_fuel_compounded import *
 
 
 def test_calculate_fuel_from_mass_compounded():
